+++ main_program/Library/payment_framework/CHANGELOG.md
@@ -1,4 +1,4 @@
 # CHANGELOG
 ## [1.0.0] -2025-09-19
-### Add
--new started!
+### Added
+- Added `pay_money` function to handle customer payments and update balance data, if payment successfully return True else return False